🔍 Yangi turdagi ma'lumotlar qo'shildi, bu kodni to'liqroq va qulayroq qilishga yordam beradi.
