+++ internal/runtime/types/int.go
@@ -1,0 +1,12 @@
+package types
+
+type IntValue struct {
+	Value int
+}
+
+func NewInt() Object {
+	return &IntValue{}
+}
+func (a *IntValue) GetValue() any {
+	return a.Value
+}
